Add text input for schema to Special:NewSchema
Since we might want to support other kinds of schemas or other ShEx
syntaxes in the future, explicitly use the name “ShExC” where possible.
(Shape Expressions Compact Syntax, or ShExC, is the most human-readable
syntax for Shape Expressions, but not the only one – they can also be
represented as JSON-LD or RDF Turtle.)

This also adds an id attribute to the schema element on the content
page, so that it can be extracted in the tests.

Bug: T214571
Change-Id: Ibe809ba72813132e8624ad37f136fef44b3622f4