Add automatic linting as pre-commit hook
This automatically runs eslint on all staged .js files. It also tries
to fix those files and adds the fixed code the commit. Similarly for
.json files and styles.

This handles partially staged files correctly.

It also works for .php, but due to non-standard phpcbf exit codes we
cannot use git add there 🙄

I'm not sure how to use this with banana.

Change-Id: I891d5e16e302826b77d39f4c29b62a0f00212570