Use waitForExist() instead of waitForVisible()
The label, description, aliases, and ShExC of a schema can all be
completely empty (the label is required when creating a new schema, but
not when editing it), so to ensure that we can always locate them on the
page, use waitForExist() (which supports empty elements) instead of
waitForVisible() (which does not).

This also makes us a tiny bit more compatible with WebdriverIO v5+,
where there is no more waitForVisible() method.

Change-Id: I10571a018953b01fe3b68cff8d45097867b44e2c