Disable creating Schema via Edit Action
This disables creating a Schema via the ?action=edit page.
It also overwrites the text of an empty page in the Schema namespace to
point to the NewSchema special page for creating a new Schema.

Later, we might also have to check for redirects, but, currently,
creating a redirect doesn't seem to be possible.

Bug: T214912
Change-Id: Ifb211c60c2c7a3325773dce794285ca2d21ee3a7