Refactor to use consistent return
This makes some minor refactorings, which I actually should have noted in
the code review before giving +2:
* consistent return: either always return a bool or never, but since it
seems to be not necessary to return true, I removed it.
* Objects are always passed by reference: Keeping that reference might
be necessary to keep conform with a legacy signature that is overwritten
or implemented. But since this is just a callback, that is not necessary
here.
* Always use strict comparison, unless comparison with type conversion
is necessary. Though, I'm not sure if $title->getLength() is really the
best method, it seems to work.

This commit is a follow-up to: I140f544a7c7a657ddc6a0aac922a59ca3f455a19

Change-Id: I8e0c950a83061f3a8af2749ee93d8021be002e0a