Inline SchemaEditAction::getEditOrViewAction()
A method that can return different kinds of action already doesn’t
really belong in any particular action class, in my opinion, and this
becomes more significant as I want to add more different classes here
(for “undo”, which is not really a FormAction).

Also use Action::factory() to construct the “view” action instead of
hard-coding the ViewAction class, so that we don’t have to change this
handler if we override that action in the future.

Change-Id: I4913beb1d380663c93a03e077c11822ae46b7e7e