Add SchemaRepository::loadSchema()
This is the counterpart of SchemaRepository::storeSchema(), allowing you
to load a schema by its ID.

Since the schema ID is not currently part of the serialization,
RevisionSchemaRepository::loadSchema() assigns the ID of the Schema
object after deserialization. I suspect we’ll want to improve this in
the future, but for now it should be good enough.

Change-Id: Ic74a8aec0cef17a18c6126e328150fe786c69a45