Add WikibaseSchemaSlotDiffRenderer
This renders a Diff between two schemas into rows for an HTML table,
based on the usual MediaWiki diff format. Additionally, when used as a
SlotDiffRenderer it also performs the diffing itself, and the content
handler is updated to register it as the diff renderer for schema
content. I chose to do the rendering and the SlotDiffRenderer
implementation in one class, instead of splitting it across two, for two
reasons: first, the MediaWiki SlotDiffRenderer abstract class is fairly
small and does not include a lot of baggage, and second, the output of
the renderer is fairly MediaWiki-specific anyways (it expects to be part
of a table where the first and third columns are much less wide than the
second and fourth columns, etc.), so a separate renderer class would not
be very useful on its own.

Bug: T214396
Change-Id: I10ce3b1db56baed0d6b6f37fa73b56cc9209b7e7