Add Schema patcher
Apply a diff to a Schema and return the result.
Throw an exception if the diff cannot be applied.

Irritatingly, the null coalescing operate doesn't work on ArrayObjects
in hhvm. It does just fine in PHP7 🙄
That is the reason that the construct
isset( $patch['labels'] ) ? $patch['labels'] : null
is used in this patch.

Bug: T214914
Change-Id: I3f1cfaee5075e1df2298eb2b273f39d79690b4e4