Add SchemaWriter::updateSchemaContent()
This function is used to just replace the schema content (or schema
text), while leaving the labels, descriptions and aliases untouched.

No custom edit summary is used yet, since that will change soon anyways
due to T214887, and I don’t want to make the translators work on a
message that will soon be removed or changed again.

The different serialization versions aren’t really dealt with (that
would probably require some integration with SchemaDispatcher); instead,
we make use of the fact that the schema text is stored the same way in
both versions, and just update that while leaving the rest of the
serialization intact. This does mean that we may write new revisions
with old serialization versions, but for the time being that seems
acceptable to me, since we need to be able to read those serializations
anyways for historic revisions.

Edit conflict protection is not properly implemented yet: the schema
text is overwritten with no respect to the previous content. When used
interactively, this is the wrong thing to do: part of the request data
should be the revision that the user’s input (the new schema text) was
based on, and we should check PageUpdater::hasEditConflict() to see if
any changes have been made since then, and possibly throw an error in
that case (or handle the changes – for example, if only labels,
descriptions or aliases were edited since then, there is nothing to do,
and we could attempt to merge changes in the schema text). PageUpdater’s
built-in protection only guards against modifications between
grabParentRevision() and saveRevision(), which is a very short time
frame and not the one where conflicts are likely to happen.

Based on I02adffc5da by Thomas Arrow <thomas.arrow_ext@wikimedia.de>.

Bug: T215395
Change-Id: I897c221241bbef9575ab427867bcfb5d7e8f84e4