Show all languages when viewing schemas
This is just a rudimentary patch to load and show all name badges when
viewing a schema, instead of just the English one, to make development
of *editing* schemas in other languages (T216145) easier. The output
will have to be improved later – currently, for example, it creates
multiple elements with the same id attribute, and also takes up a lot of
screen space if there are many languages.

Currently, the user language (according to the parser options) is used
for the “primary” language (the first name badge to show), which causes
the parser cache to be split by the user language. We need to keep in
mind that this is not necessarily the same as the user language
according to the $wgLang global (this has caused issues in Wikibase in
the past, see T199983 and T201170), and watch out for cache behavior if
we make other page elements depend on the user language (e. g. when
using interface messages).

One of the test cases in WikibaseSchemaContentTest has to be abbreviated
a bit to still fit in the line length limit.

Bug: T216144
Change-Id: If8fa7de6b85185dd03c5c3fb094c43b506926356