Properly clean up database around tests using it
When we don’t declare the database group, or more specifically which
tables in the database we touch, then the tables may only be cleaned up
at the beginning and end of the whole test run, and the different tests
that create revisions, often for the same page (Schema:O123), all end up
seeing each other’s data. That this has worked out so far is, as far as
I can tell, sheer luck, and it broke down when I was changing the
serialization format from 'schema' to 'schemaText' and tests were
overwriting each other’s data with different serialization versions.

The list of tables to reset is probably incomplete, but it seems to be
good enough for now.

(A lot of these tests should really not be using the database at all,
and instead access some mock revision store, but fixing that is a larger
undertaking that I don’t have the time for right now.)

Change-Id: I7832e45d66bba44592dd068067a184d2fb874329