Move and split RenderDiffHelper
The RenderDiffHelper is not really a service, but core part of the
Presentation and thus can live well in a Presentation directory.
This also splits the class up, since the diff and the confimation form
are two separate parts of the page, even though they are currently used
in the same context.

Change-Id: I5626fdd33510769719d3367a9a3f108e539b60dd