Add link to check entities against schema
The link is only added if the target is configured, which is not the
case by default. In production, this is intended to point to
https://tools.wmflabs.org/shex-simple/wikidata/doc/shex-simple.html, but
that service is not suitable for all installations (an HTTPS page cannot
load the schema text from HTTP wikis, and Toolforge also has a Content
Security Policy, though it’s currently in report-only mode).

We might also want to add an icon to the link (not sure).

Bug: T217331
Change-Id: I6f55df112360435877a36e0d6dd5c812fbda6ec7