Extract WikibaseSchemaSlotViewRenderer
This class, extracted from WikibaseSchemaContent, takes care of
rendering a FullViewSchemaData in a certain language into a
ParserOutput. Making the language a class member allows us to remove a
$languageCode from several methods, which was inconvenient and confusing
(since several other methods have an unrelated $langaugeCode argument,
which is still there). This is also a good opportunity to allow for
proper dependency injection of the LinkRenderer (though we currently
don’t actually inject it anywhere).

The existing tests for WikibaseSchemaContent are all changed to tests
for WikibaseSchemaSlotViewRenderer, since there were no tests for any
other parts of WikibaseSchemaContent. However, WikibaseSchemaContent
gains new tests, checking that it uses the correct language from the
parser options and also splits the parser cache by it (which happens
automatically in ParserOptions, but just to be safe).

Bug: T216144
Change-Id: I305367ea861820ede8a5478cfc81b951892f86f2