Introduce SpecificLanguageMessageLocalizer
Instead of storing a language code in WikibaseSchemaSlotViewRenderer,
using the global wfMessage() function to create each message, and having
to remember to override its language every time, it’s better to create a
new implementation of MediaWiki’s MessageLocalizer interface that uses a
specific language code for all messages, and use that as a service to
create all messages. We can also pass this service into other places in
the future if it becomes necessary, to ensure that they, too, use the
right language.

Bug: T216144
Change-Id: Ic73e1f1299c431d01224deebfc5a7b13cc6dc684