Report edit conflicts when editing name badge
Add a hidden input to Special:SetSchemaLabelDescriptionAliases, which
records the base revision ID of the edit – the ID of the revision from
which the name badge was originally loaded, upon which the user’s edits
are based. If this is no longer the latest revision at the time the edit
is saved, an error is reported.

Bug: T218298
Change-Id: I297267e9c8841789ec2c89e159a6e2697b17059e