Use LinkRenderer::makeKnownLink instead of ::makeLink
All the links we render here *should* always be known (same page as the
current schema, or a special page), but MediaWiki apparently tries to
render new pages before they are fully saved, so the link appears as
“unknown” for the initial save of a newly created schema (until the
parser cache is purged). We can circumvent this issue by informing the
LinkRenderer that these links are known, so that it doesn’t have to rely
on Title::exists().

Bug: T218769
Change-Id: I55987180e678f03916c9b0954fac0188e7eb974a