Add browser tests for edit conflict detection
Add browser tests checking that, when the same form to edit a schema’s
name badge or schema text is submitted again, an edit conflict is
reported and no change is made.

Note that MediaWiki core suppresses edit conflicts between revisions by
the same user, and the case we test here (hitting “back” to re-submit
the same form) is specifically given as the rationale in T30720#321100.
However, we don’t do any such conflict suppression in WikibaseSchema.

Bug: T217338
Change-Id: Ib298720219f07244be186799d54883ec231d9526