Fix external link style
The CSS selector used by various skins to style external links is some
variation of `.mw-parser-output .external`, and since our HTML is not
output from the parser, the external link to the tool to check entities
against a Schema was not styled as an external link. Fix this by
wrapping the link in a <span class="mw-parser-output"> – this feels
safer than wrapping our whole output with such a class, because we don’t
know what other assumptions exist about parser output that aren’t true
for our output.

Bug: T218885
Change-Id: Ie4a07a89e3b69bf70b812ecca7f3f4a1cf5d1272