Use autocomments for translatable edit summaries
Since MediaWiki is not yet ready for Messages in CommentStoreComments,
use an approach similar to Wikibase: generate a special autocomment and
then use the FormatAutocomments hook to translate the message there.

Unfortunately, this means we need a new summary message, because we
can’t pass the label as a message argument in this case (the rest of the
comment is not available to the hook). The old message is kept for now,
in case we can use it again in the future, but marked {{notranslate}} to
avoid wasting translators’ time.

To make a future migration easier, we include as much information as we
can in the comment data. The comment_data field is a BLOB of up to 65535
bytes, so to make sure that the schema text fits (a long series of line
break characters will need two JSON bytes per schema text character), we
truncate it. The other fields are short enough that we can include them
without truncation at this level.

Bug: T218893
Change-Id: I5747a4328106a62a15b6f11da49012a33f94be03