Specify flags for PageUpdater::saveRevision()
EDIT_NEW signals that we intend this edit to create a new page and an
error should be raised if the page already exists; EDIT_UPDATE does the
opposite. We should specify one or the other, since we always know
whether we intend to create or update a page.

EDIT_INTERNAL means that “the page retrieve/save cycle happened entirely
in this request” and mainly seems to be used to skip the stashed
revisions mechanism (which we don’t use).

Change-Id: If95c41869b34e3f5ec1df6b68bf7b2e86ea621f6