Extract AutocommentFormatter to separate responsibilities
This extracts the parsing and formatting into a new class to reduce the
complexity of the hook method and focus it on deciding whether we want
to handle that event and then handle our autocomments in a different
class.

We now don't need that many tests for the hook anymore, as much of that
functionality is already tested for the AutocommentFormatter.

Change-Id: Ice002cf7b8213fa3d407bf2c11bf8f25742101d0