Remove unused message
We currently don’t use this message, and likely won’t use it for a few
more months (though we would like to return to a summary/comment system
that allows us to use this message again eventually). Let’s remove it
from the i18n files for now – translations on translatewiki are kept, so
if we reintroduce the message later, the translations will be added back
again, with no need for translators to re-translate them.

Change-Id: I2ac09673be8362c828b73d1354c268cf5fc9e4eb