Check for internal edit conflicts in undo/restore
This adds a $baseRevId argument to SchemaWriter::overwriteWholeSchema(),
raising an error if this is not the latest revision ID. The undo action
sets this to the revision ID against which the diff was applied, and the
restore action sets it to the revision ID provided by the user. In both
cases, the IDs are expected to be the same almost all the time (undo
loads the latest revision, and restore already compares against the
latest revision before calling SchemaWriter); only a well-timed edit
during processing, or replication lag, could trigger this error.

Bug: T219531
Change-Id: Iddbbbb29da2b5ec721e58cc0c3f7263f96868acd