Only prevent actually conflicting edits
A change to a name badge in one language shouldn't stop the saving of
an edit on a name badge in another language. Neither should a change
to the Schema text block a change to a name badge or vice versa.

However, changes within the schema text will always be considered to be
conflicting, even if a 3-way-merge would work. The reason is that the
Schema text is code and a change in one line may make a change in another
line at the other end of the schema non-sensical.

Bug: T218300
Change-Id: Ic5f4f04a3d10d0a7f305e847fdc8131c44c94c66