Split SchemaWriter: SchemaInserter and SchemaUpdater
This splits up the MediaWikiRevisionSchemaWriter class because of the
diverging functionality of the codes that initially inserts a new
Schema and the code that updates an existing Schema.
Especially, the IdGenerator is only required by the new SchemaInserter.
On the other hand, Only the new SchemaUpdater has to check for
edit conflicts.

Change-Id: I4f9d168aa3d6012de6508bd8baf96db6ebe90ea5