Fully prevent empty edits
This replaces the EditConflictDetector with a more general way of
conflict resolution: we apply the user’s edits to the base revision,
diff the resulting data to the base revision, and then patch the parent
revision with that diff. If the patch fails, an edit conflict is
reported; otherwise, the edit can be saved. This allows merging some
additional non-conflicting changes, e. g. when the label and the
description in the same language are changed simultaneously, or when
unrelated areas of the schema text are edited.

A special benefit of this is that if the diff is empty, we can skip
saving a new revision. Most of the time, the PageUpdater would otherwise
have skipped the new revision because the underlying serialization
didn’t change (we try to ensure a canonical JSON representation), but
it’s still possible that the serialization changes, e. g. because we
forgot to normalize something or because the parent revision was saved
under an older serialization version.

overwriteWholeSchema() remains unchanged, we don’t even want to try
merging anything here.

For the tests, MediaWikiRevisionSchemaUpdaterTest now has two member
variables to share the base and parent revision record across different
helper methods, and a new RevisionLookup implementation is added that
can return those revision records without requiring database access.

Bug: T219173
Change-Id: I6c7f358f30e38aaf6d73aa898872670987354051