Small maintenance script adjustments
This could have been multiple patches, but since we have to backport
these, they are now combined in this single patch.

The files should be in maintenance namespace to prevent confusion in the
future.

If MW_INSTALL_PATH is not set, go the correct number of directories up.

Since we are never setting the aliases and the Schematext, we can be
explicit about that.

Adjusted one label in one of the Schemas that are to be created.

Change-Id: I9765e8f9ebe5b9a043002a35da62afb1855376a6
(cherry picked from commit 8546e378d84dc7df3fb1d131d75f6cd8d8be065f)