Fix Exception by forcing default diff engine and renderer
The MobileFrontend extension instantiates its own InlineDifferenceEngine
that leads to triggering the deprecated
DifferenceEngineSlotDiffRenderer. The fix for this is to provide our own
fresh TextSlotDiffRenderer.

Further, in violation of LSP, this InlineDifferenceEngine does not
expect a diff body made of table rows. Thus, we also have to overwrite
the createDifferenceEngine method to force the default engine being
used.

Bug: T224545
Change-Id: I703fdef8b9aa35acb1a6eab7c8321e7fe7e271ea