Replace usage of deprecated Page in favor of WikiPage/Article
Calling Action::factory and Action::__construct with WikiPage
was soft deprecated in core favor of calling it with Article.

This is the first step in eliminating Page interface from the
extension. Once calls to Action::factory with WikiPage are
eliminated from the codebase and is deprecated, we would be
able to deprecate calling action override callback with WikiPage,
and create a followup to remove the WikiPage->Article conversion
logic from this extension.

Bug: T239975
Change-Id: I2956b17079b0d233ed672905ea00355bad5b322f