Fix calls to OO.ui.infuse
Passing a string ID was removed a year ago.

This also requires some changes to the browser tests. When OOUI replaces
the infused elements, WebdriverIO shouldn’t keep waiting for the old
elements (no longer attached to the DOM) to become displayed – instead,
we should re-run the selectors on each iteration.

Bug: T269167
Change-Id: I1393453d50f171f64abc50649596d316036a3612