Fix EntitySchemaSlotDiffRendererTest for diff marker change
MediaWiki core changed the way diff markers are generated in change
I926bedf715 (commit 27319def63, T192526). Update the expected test data
to match this. At some point, we should also update our own diffing code
to do the same thing, but that’s not as urgent as unbreaking the tests.

Bug: T272329
Change-Id: I05e92b80081480b5d7213cd0123251bf0dd6f25b