SECURITY: Replace more block checks permission checks
Similarly to Iac86cf63bd, update other places which explicitly checked
if the user is blocked, and get all possible kinds of permission errors
instead. This also subsumes separate calls to userHasRight().

In NewEntitySchema, we inherit checkPermissions() from SpecialPage, but
that method doesn’t seem to do all we need (it boils down to just a
userHasRight() check), so we add a new checkPermissionsWithSubpage()
method that gets all the permission errors.

The browser tests need some slight adjustments, since the blocked error
on Special:NewEntitySchema looks a bit different now.

Bug: T296578
Change-Id: Id9af124427bcd1e85301d2140a38bf47bbc5622c
Signed-off-by: toan <tobias.andersson@wikimedia.de>