tests: Migrate user blocked tests to Cypress
The original Selenium/webdriver tests blocked the Admin user for the
duration of the test, and unblocked that user at the end of it, in an
`afterEach` hook.
However, Cypress uses a different paradigm that focuses more
debugability and strongly discourages the use of `after` and `afterEach`
hooks, so that the state of a test is preserved at its end.
(See
https://docs.cypress.io/guides/references/best-practices#Using-after-Or-afterEach-Hooks)

This would either mean to unblock the Admin user in a beforeEach hook,
or to use block a dedicated or newly created user. The drawbacks of the
first option by far outweigh its usefulness, so focusing on the second.

Making the API calls to create the new users directly in the test
itself, maybe similar to how we use pageobjects, won't work well,
because Cypress tests run in the browser itself, so any call to
mwbot.login() would log the current user in, that is, we cannot handle
multiple self-contained API users. With this approach, we would have to
carefully track at any point which user is currently logged in,
especially when a user is supposed to be only get blocked in the middle
of a multi-step process.

Therefore, this change introduces a Cypress plugin for Mediawiki API
calls to create a new user. With this plugin we can make the API
requests from node and are not interfering with the browser session. It
uses the api-testing library, because that library already provides use
with almost all the functionality we need.

Future improvements could be:
* storing/reading users from the state and making sure they're unblocked
  when requested again
* reading users + passwords from env variables to skip user creation
* extract this into its own npm library (and rewrite in typescript)
* create a similar plugin for Wikibase commands, for example for
  creating an Item, or Property (and being env-aware for the properties)

Change-Id: I9a6e09bffe645b1284693907f1f12cef59adf654