Set up initial validators for entity-schema data type
Validate that the value looks syntactically valid. We don’t yet check
that the referenced EntitySchema actually exists, that will be done
separately.

The maximum string length is chosen based on Int32EntityId::MAX (though
that class is not actually related to SchemaId): ten decimal digits
(enough to fit a 32-bit integer), plus one character for the ‘E’.

Bug: T333813
Change-Id: I89c9288fdf447c3a7da6d6abf8c7114fefe38402