Fix phpcbf exit code for lint-staged
phpcbf exits with a failure exit code if there were any violations, even
if they could all be fixed. This makes it useless in a lint-staged
pipeline (which expects scripts to exit successfully, and will otherwise
discard all changes and abort the commit), so wrap phpcbf in a small
shell script checking the exit code and ignoring a 1 exit code. (Exit
code 2 would mean that there were some errors that could not be fixed.)

We can hopefully revert this when PHP_CodeSniffer version 4 is released,
which is supposed to change the exit codes to make them more sensible
(squizlabs/PHP_CodeSniffer#2898).

Change-Id: Ida87fd259bd7cc1b6c0f789001a574865e435472