Update eslint to allow es2018 and require trailing commas
This references the commit that is currently at the head of master of
https://github.com/wikimedia/eslint-config-wikimedia in order to already
pull in the updated browserlist-config-wikimedia. Otherwise, it would
complain about Promise not being supported in IE11. This workaround will
be rectified in one of our upcoming chores, as soon as the new version
is released. This approach seems preferable over disabling the
compat/compat rule, because we might forget to reenable it again after
updating eslint-config-wikimedia.

Bug: T332554
Change-Id: I71cf2b201ac37e8a2e6d1e84bc96c69e61b5faa8