Remove documentation for Selenium from README.md
The last bits of webdriver selenium tests have been removed in I7cdc0576
(commit 02f7407). This change now cleans up the outdated parts of the
README. The browser-test section should now only be about Cypress.

However, CI still invokes `npm run selenium-test` to execute the browser
tests. So we need to have this command available and it makes sense to
also document it in the README.

Bug: T333429
Change-Id: I08a81aaac3c77ad7dbb302209a08ad708b997845