Add LabelLookup service
We want to show EntitySchema Entities/pages in a lot of contexts not
(only) by their id, but also by their label. Thus we need a reusable
LabelLookup service.

This is a simple implementation of such a service. It loads the whole
EntitySchema Entity and picks the label if it exists in the given
language. It intentionally does not apply any fallbacks yet, that will
come later in T330491.

The current implementation is somewhat inefficient by loading the whole
entity when we only want a label, but we can introduce
performance optimisations if and when that becomes actually an issue.

Bug: T338613
Bug: T339924
Change-Id: I82aec84693b9471b9d2ed4c43cb9bda514ed6b14