Show EntitySchema Labels on special pages instead of only the id
This is a trimmed-down version of the same class in Wikibase.

We have to inject the whole RequestContext object, because accessing its
language in the factory causes problems with Visual Editor. Apparently
some VE RL module involves parsing a message with a link. That means
that the hook handler is called from load.php, where we're not allowed
to look up the language.

Bug: T343231
Bug: T339921
Bug: T339925
Bug: T339924
Depends-On: Ic1d8b39b053fb0f6050bd7cbdfe7b140dd423659
Change-Id: I6dbb683d755b1a4eed6f0b158b26185cf11dbbde