Show copyright notice and anon-user warning when editing schema text
We should inform the user about the license everywhere they
edit in the UI. And also, all those place should inform users that are
not logged in that their IP address will be recorded. Those two messages
were missing from the SchemaEditAction, which was an oversight back from
when originally creating that.

This change now adds the mechanism, the same way it is added for the
special pages to create a new EntitySchema and to edit the Labels,
Descriptions and Aliases.

Bug: T343333
Bug: T343118
Change-Id: I5554465010a46af8d12c2fe32724955c4b624837