Improve HTML formatting in HtmlPageLinkRendererEndHookHandler
Previously, we were only emitting a subset of the HTML emitted by
Wikibase’s DefaultEntityLinkFormatter, despite reusing its messages; in
particular, we weren’t language-tagging the HTML according to the
language of the label being displayed, which we really should do once we
have language fallback.

The old code also wasn’t handling HTML-escaping quite properly, though
it still ended up safe – ->rawParams()->escaped() results in unsafe
HTML, but we returned it as string not wrapped in HtmlArmor, so the
LinkRenderer escaped it for us again. (The only practical effect was
that, if the messages themselves contained any HTML, it would now be
double-escaped.) Instead, we now return proper HtmlArmor (so that the
real HTML markup we add, with the <span>s, isn’t escaped), and the
HtmlArmor::getHtml( $label->getText() ) ensures that the label text is
HTML-escaped.

Bug: T338798
Change-Id: Ifa5ccb9aa9cc025ac557eb1bcdd4b5a475e19c0f