tests: Extract traits for unit and integration tests
There was a lot of copy+pasted code in unit and integration tests; move
to some common traits: EntitySchemaUnitTestCaseTrait for unit tests, and
EntitySchemaIntegrationTestCaseTrait for integration tests.

At least EntitySchemaIntegrationTestCaseTrait needs to be a trait and
not just a parent class, so that the API and special page tests can
extend their more specific parent classes. We could *additionally* have
a parent class EntitySchemaIntegrationTestCase that just extends
MediaWikiIntegrationTestCase and uses the trait; I wouldn’t mind that,
but right now I’m also not sure if it’s worth it or not.

Some of the saveSchemaPageContent() implementations were returning the
revision ID (int) instead of the RevisionRecord, so those call sites
have to be changed to add a ->getId() call. But I think returning the
whole RevisionRecord is nicer (more self-documenting).

In EntitySchemaFormatterTest, we can just remove the setup method – in
an integration test, the extension.json constants are guaranteed to be
loaded. (The setup method used to be necessary before the test was
turned from a unit test into an integration test in change I2fa7577318,
commit 210685ac2d).

Change-Id: I0dd0a5085f36f052e46faa3bd90a693d0cd1dbb1