Make EntitySchemaValue::getArrayValue() match EntityIdValue
I added 'type' here in one of the patch sets of change Ie3d4350ed0
(eventually merged as cecd1c47f7), apparently quite carelessly without
checking the corresponding Wikibase code, because in Wikibase it’s
actually 'entity-type'. Fix that now, while we still can (as of this
writing, as far as I can tell, “new-style” EntitySchema values, with
entity ID values rather than strings, only exist on Beta but not on Test
Wikidata).

Also use this opportunity to change the type from 'entityschema' to
hyphenated 'entity-schema'. This makes it match the data type, and while
the two don’t have to match (e.g. data type 'wikibase-item' ≠ entity
type 'item'), it seems less confusing and slightly more readable to use
the hyphenated form for both if we can; as far as I know, there’s no
technical reason in Wikibase to avoid hyphens in entity type names. (We
don’t register 'entity-schema' as a real entity type yet, but once we
do, I assume we’ll want that identifier to match the 'entity-type' in
the data value serialization, so I think this is the last point where
we’re still free to change it – once we start to have statements with
'entity-type' => 'entity-schema' in Test Wikidata, we’ll want to avoid
breaking those.)

Bug: T362955
Bug: T362001
Change-Id: I4fccf2285e2267805ef63b60d99daf7252c7c432