Add weekly keepalive / heathcheck mail for secondary CI
Secondary CI jobs run on Github and send mails to alert of CI
failures. When the GMail credentials for sending the mails have
not been used in a while, Google expires the credentials.

Send a regular weekly e-mail to keep the credentials active. The
target address is a mailing list with no members.

Bug: T365704
Change-Id: I76f34817f2bd72f93d88d8a3f2bb5c0ecd9f2285