Only register EntitySchema namespace when feature is enabled
We want clients and repos to be able to use the EntitySchema
extension and process items with EntitySchema-typed statements,
but we don't want Wiki clients with the extension enabled to be
able to create pages in the EntitySchema namespace.

Make the registration of the namespace conditional on the wiki
being a repo, and the `EntitySchemaIsRepo` config being set to
true.

Bug: T363153
Change-Id: Ia52f6c3d9340daf4422b241ed264d7405c95f24d