Check EntitySchemaIsRepo in more hook handlers
All the title-related hook handlers are very unlikely to make a
difference: but this in theory makes it more possible to run
EntitySchema in client-only mode on a wiki where namespace 640 is an
unrelated namespace.

LoadExtensionSchemaUpdatesHookHandler is pretty important, though –
without this, we were creating the entityschema_id_counter table on
client wikis, where it’s not used and shouldn’t exist. (This doesn’t
apply in production, where we never run update.php [1], but I could
confirm that the table was created on Beta testwiki.)

Note that LoadExtensionSchemaUpdatesHookHandler is not allowed to have
services injected, so we get the config variable as a global instead.
Additionally, calling WikibaseSettings::isRepoEnabled() doesn’t work for
some reason (“Error: Class 'Wikibase\Lib\WikibaseSettings' not found”,
at least in CI), so inline it and directly check the ExtensionRegistry
instead. (This also means its test needs to become an integration test.)

[1]: https://bash.toolforge.org/quip/TNR_cHEBv7KcG9M-Utqd

Bug: T363153
Change-Id: Iaae48b794bfe7ff54d4e209bf64c13ff6f2a654c
(cherry picked from commit 871305b6cf0cf9dd40a9e270a02bf8329d314326)