Improve namespace registration
The previous way to register the namespace conditionally was partially
broken: the assignments to $wgNamespacesWithSubpages,
$wgContentNamespaces and $wgNamespaceContentModels had no effect,
because by the time the CanonicalNamespaces hook runs, the NamespaceInfo
service has already been initialized with copies of these variables, and
subsequent changes to the globals are not seen by the service anymore.

There is a much better way: still register the namespace in
extension.json, but with "conditional": true, which excludes the
namepace from the default list of canonical namespaces (but still sets
the other globals: it’s fine if they contain data for “unused” /
disabled namespaces). Then, in our CanonicalNamespaces hook handler, we
only need to add the namespace there if it’s supposed to be enabled, and
don’t have to do anything else. This was previously poorly documented at
[1] (but mentioned at [2]); I’ve improved it now [3][4].

This has the theoretical downside that it’s no longer possible to have
EntitySchema in client-only mode alongside an unrelated extension that
also happens to register a namespace with the ID 640 without issues
(EntitySchema would now declare that unrelated namespace to be content,
not movable, etc.) – but I don’t think that’s generally expected to be
supported (extensions are supposed to look at [5] when picking their
namespace numbers and avoid stepping on each other’s feet that way, so
there should be no other extension with namespace 640).

[1]: https://www.mediawiki.org/w/index.php?title=Manual:Using_custom_namespaces&oldid=6542395#In_extensions
[2]: https://www.mediawiki.org/wiki/Manual:Extension.json/Schema#namespaces
[3]: https://www.mediawiki.org/wiki/Manual:Using_custom_namespaces#In_extensions
[4]: https://www.mediawiki.org/w/index.php?title=Manual:Using_custom_namespaces&diff=prev&oldid=6606006
[5]: https://www.mediawiki.org/wiki/Extension_default_namespaces

Bug: T368010
Change-Id: I02fe6a386662a7524de3c96a923a2ae9718cd0c2