Tag EntitySchema client integration tests with EntitySchemaClient
EntitySchema can be enabled for client wikis. We have some
integration tests for functionality that we would expect also to
work in a client-only configuration.

In I93a08d9e8c88aacd2ec4dce8a052343fbf365a13, we enable the
Wikibase client-only jobs for EntitySchema CI checks. With that
change, only tests tagged with group `EntitySchemaClient` will be
run during the job. Add that tag to the tests that we want to be
included in the EntitySchema client-only CI checks.

Bug: T367156
Change-Id: I545fd5a61157d8a8da8877dda80a05d7d0ea4ccd