Add error handling for HookRunner failures during EntitySchema create
The NewEntitySchema special page currently throws a RuntimeException
if any error is encountered running the `EditFilterMergedContent`
hook on schema creation. The exception bubbles up to the top-level
and is presented as a stack trace in Mediawiki. This can, for
example, be the case when an action is blocked by the `AbuseFilter`
plugin.

Create a new `HookRunnerFailureException`, and throw it in case the
hook fails. Catch the exception in the special page and return the
failed status from the HTMLForm's submit callback so that OOUI can
present it is a meaningful form submit failure.

Bug: T365452
Change-Id: I4c6d35c24530cd8d99ac3900ec5f718f351fc0c8