Limit width of `mw-highlight` div in schema view with grid layout
After some recent changes to SyntaxHighlight_GeSHi
(Ic58482b73e520505bde26a47e74f131dd8a51671) to support dark mode,
the background colour of the `mw-highlight` div on EntitySchema
pages became different to the default (`inherit`ed) background
colour. This made it apparent that the background of the
`mw-highlight` fills the complete page width -
`entityschema-schema-text-links` is floating to the right, on top
of the highlight div. This is not what we want - we want the links
section to have the background colour of the main page even if the
schema section has a different colour.

Change the CSS of the EntitySchema view page to put the Schema and
Links divs in a grid. This gives the Links div a real width and
allows it to preserve the `inherit`ed background colour, leaving
the Schema div to behave as it chooses.

Bug: T369283
Change-Id: I586731772c6060ee9a7515a86fbd7b8d624f9873